Restructure packaging for qr code
COAND-497
