+++ action/src/test/java/com/adyen/checkout/action/TestActionDelegate.kt
@@ -32,7 +32,7 @@ import com.adyen.checkout.components.ui.ViewProvidingDelegate
 import com.adyen.checkout.components.ui.view.ComponentViewType
 import com.adyen.checkout.core.api.Environment
 import com.adyen.checkout.core.exception.CheckoutException
-import com.adyen.checkout.qrcode.QRCodeOutputData
+import com.adyen.checkout.qrcode.internal.ui.model.QRCodeOutputData
 import com.adyen.threeds2.customization.UiCustomization
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
