+++ qr-code/src/main/java/com/adyen/checkout/qrcode/internal/ui/QRCodeDelegate.kt
@@ -6,16 +6,20 @@
  * Created by oscars on 12/8/2022.
  */
 
-package com.adyen.checkout.qrcode
+package com.adyen.checkout.qrcode.internal.ui
 
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.base.ActionDelegate
 import com.adyen.checkout.components.base.DetailsEmittingDelegate
 import com.adyen.checkout.components.base.IntentHandlingDelegate
 import com.adyen.checkout.components.base.StatusPollingDelegate
 import com.adyen.checkout.components.base.ViewableDelegate
 import com.adyen.checkout.components.ui.ViewProvidingDelegate
+import com.adyen.checkout.qrcode.internal.ui.model.QRCodeOutputData
+import com.adyen.checkout.qrcode.internal.ui.model.QrCodeUIEvent
 import kotlinx.coroutines.flow.Flow
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 interface QRCodeDelegate :
     ActionDelegate,
     DetailsEmittingDelegate,
