+++ qr-code/src/main/java/com/adyen/checkout/qrcode/internal/ui/model/QRCodeOutputData.kt
@@ -6,10 +6,12 @@
  * Created by josephj on 13/4/2021.
  */
 
-package com.adyen.checkout.qrcode
+package com.adyen.checkout.qrcode.internal.ui.model
 
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.base.OutputData
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class QRCodeOutputData(
     override val isValid: Boolean,
     val paymentMethodType: String?,
