+++ qr-code/src/test/java/com/adyen/checkout/qrcode/QRCodeComponentTest.kt
@@ -17,6 +17,8 @@ import com.adyen.checkout.components.ActionComponentEvent
 import com.adyen.checkout.components.base.ActionComponentEventHandler
 import com.adyen.checkout.components.model.payments.response.QrCodeAction
 import com.adyen.checkout.core.log.Logger
+import com.adyen.checkout.qrcode.internal.ui.QRCodeDelegate
+import com.adyen.checkout.qrcode.internal.ui.QrCodeComponentViewType
 import com.adyen.checkout.test.TestDispatcherExtension
 import com.adyen.checkout.test.extensions.invokeOnCleared
 import kotlinx.coroutines.ExperimentalCoroutinesApi
